pyproject.toml
[project]
name = "diffusion_quantum_demo"
version = "0.1.0"
description = "Quantum diffusion demo environment"
requires-python = ">=3.9"
dependencies = [
  "streamlit",
  "py3Dmol",
  "rdkit-pypi",
  "numpy",
  "matplotlib",
  "qiskit",
  "qiskit-nature",
  "pyscf",
  "torch",
]

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/diffusion_quantum_demo"]


